+++ src/Middleware/Router.php
@@ -10,23 +10,28 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Yiisoft\Http\Header;
 use Yiisoft\Http\Method;
 use Yiisoft\Http\Status;
 use Yiisoft\Middleware\Dispatcher\MiddlewareDispatcher;
 use Yiisoft\Middleware\Dispatcher\MiddlewareFactory;
 use Yiisoft\Router\CurrentRoute;
+use Yiisoft\Router\MethodFailureHandlerInterface;
 use Yiisoft\Router\UrlMatcherInterface;
 
 final class Router implements MiddlewareInterface
 {
     private readonly MiddlewareDispatcher $dispatcher;
+    private bool $ignoreMethodFailureHandler = false;
 
     public function __construct(
         private readonly UrlMatcherInterface $matcher,
         private readonly ResponseFactoryInterface $responseFactory,
         MiddlewareFactory $middlewareFactory,
         private readonly CurrentRoute $currentRoute,
-        ?EventDispatcherInterface $eventDispatcher = null
+        ?EventDispatcherInterface $eventDispatcher = null,
+        private readonly ?MethodFailureHandlerInterface $allowedMethodsHandler = null,
+        private readonly ?MethodFailureHandlerInterface $methodNotAllowedHandler = null
     ) {
         $this->dispatcher = new MiddlewareDispatcher($middlewareFactory, $eventDispatcher);
     }
@@ -37,15 +42,10 @@ final class Router implements MiddlewareInterface
 
         $this->currentRoute->setUri($request->getUri());
 
-        if ($result->isMethodFailure()) {
-            if ($request->getMethod() === Method::OPTIONS) {
-                return $this->responseFactory
-                    ->createResponse(Status::NO_CONTENT)
-                    ->withHeader('Allow', implode(', ', $result->methods()));
-            }
-            return $this->responseFactory
-                ->createResponse(Status::METHOD_NOT_ALLOWED)
-                ->withHeader('Allow', implode(', ', $result->methods()));
+        if (!$this->ignoreMethodFailureHandler && $result->isMethodFailure()) {
+            return $request->getMethod() === Method::OPTIONS
+                    ? $this->getAllowedMethodsResponse($request, $result->methods())
+                    : $this->getMethodNotAllowedResponse($request, $result->methods());
         }
 
         if (!$result->isSuccess()) {
@@ -57,5 +57,40 @@ final class Router implements MiddlewareInterface
         return $this->dispatcher
             ->withMiddlewares($result->route()->getData('enabledMiddlewares'))
             ->dispatch($request, $handler);
+    }
+
+    public function ignoreMethodFailureHandler(): self
+    {
+        $new = clone $this;
+        $new->ignoreMethodFailureHandler = true;
+        return $new;
+    }
+
+    /**
+     * @param string[] $methods
+     */
+    private function getAllowedMethodsResponse(ServerRequestInterface $request, array $methods): ResponseInterface
+    {
+        return $this->allowedMethodsHandler !== null
+                ? $this->allowedMethodsHandler
+                    ->withAllowedMethods($methods)
+                    ->handle($request)
+                : $this->responseFactory
+                    ->createResponse(Status::NO_CONTENT)
+                    ->withHeader(Header::ALLOW, implode(', ', $methods));
+    }
+
+    /**
+     * @param string[] $methods
+     */
+    private function getMethodNotAllowedResponse(ServerRequestInterface $request, array $methods): ResponseInterface
+    {
+        return $this->methodNotAllowedHandler !== null
+                ? $this->methodNotAllowedHandler
+                    ->withAllowedMethods($methods)
+                    ->handle($request)
+                : $this->responseFactory
+                    ->createResponse(Status::METHOD_NOT_ALLOWED)
+                    ->withHeader(Header::ALLOW, implode(', ', $methods));
     }
 }