+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 4.0.1 under development
 
 - Enh #265: Adapt summary data in debug collector (@rustamwin)
+- New #249: Add option to ignore method failure handler to the 'Router' middleware (@olegbaturin)
+- New #249: Add custom response handlers for the method failure responses to the 'Router' middleware (@olegbaturin)
 
 ## 4.0.0 February 25, 2025